src/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.59.4
azure-storage-blob==12.24.0
azure-identity==1.19.0
kubernetes==31.0.0
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.7.0
